pyproject.toml
[project]
name = "wecom-aibot"
version = "0.1.0"
description = "企业微信智能机器人 Python SDK - WebSocket 长连接通道"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "wecom" }]
keywords = ["wecom", "wxwork", "bot", "aibot", "websocket"]
dependencies = [
    "websockets>=12.0",
    "httpx>=0.27.0",
    "pyee>=11.0.0",
    "pycryptodome>=3.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.8.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wecom_aibot"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true
